Temporary file
Hey man, I'm totally exhausted with today, and I'm mentally kinda out of it. Tomorrow I'm gonna work out in the morning, eat something good, and I'll have this compilation errors fixed up, but here's what I've got so far. The big thing is I need to figure out how to fix the dropdown problem, cause right now it's not doing that, and I may just need to copypaste this in later when I reset the project.